Fixes from testing, refactoring of names, simplification of RulesLoader and updated documentation
